🚚 Chore(#24): move onboardingStepStore to store directory
