+++ src/app/onboarding/page.tsx
@@ -2,14 +2,13 @@
 
 import React, { use } from 'react';
 
-import useStepStore from '@/utils/onboardingStepStore';
-
 import FifthStep from '@/containers/onBoarding/FifthStep';
 import FirstStep from '@/containers/onBoarding/FirstStep';
 import FourthStep from '@/containers/onBoarding/FourthStep';
 import SecondStep from '@/containers/onBoarding/SecondStep';
 import ThirdStep from '@/containers/onBoarding/ThirdStep';
 import ZeroStep from '@/containers/onBoarding/ZeroStep';
+import useStepStore from '@/store/onboardingStepStore';
 
 import Image from 'next/image';
