+++ src/store/onboardingStepStore.ts
@@ -1,4 +1,3 @@
-import { get } from 'http';
 import create from 'zustand';
 
 interface OnboardingStepState {
